refactor: remove unnecessary input type builders
